Enjoy A Luxurious Grand Lifestyle At Prestige Neopolis Flats In Hyderabad
Prestige Neopolis is a recent project by the Prestige Group of builders in Hyderabad. This project is based in one of the centrally situated places in Hyderabad. Hyderabad has been rated as the most favourable living town in India. A lot of people are choosing to relocate to Hyderabad after retirement to live in a peaceful calm atmosphere and yet avail the of top-class advancements in all fields.

Talking about the community features, you would find some of the best high-class conveniences within the property. For instance, there is an exclusive clubhouse of various sq. ft. of area. There is a huge wide acre of Central Green Yard. There are plenty of indoor and outdoor games offered. There is a hobby centre, clinic, library, car wash area and etc. This residential project offers 1/2/3 BHK luxury apartments.

The other community conveniences fall into various categories such as cafeteria, landscaped garden, mediation centre, kids’ play area, swimming pool, exercise room, power and water back up of 24 hours, conference room, theatre, multipurpose hall, conference room, daycare centre and so on.

This property is coming up at Kokapet, Hyderabad. Prestige Neopolis Kokapet is a great position which is very close to all commercial destinations of Hyderabad. The property is very closely located in the IT hub of Hyderabad and the financial area of Hyderabad.

For More Details:
Visit Here: https://bit.ly/3znYgbu